examples/fetch-data/requirements.txt
certifi==2024.7.4
charset-normalizer==3.3.2
idna==3.7
markdown-it-py==3.0.0
mdurl==0.1.2
numpy==2.1.0
pandas==2.2.2
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
rich==13.7.1
six==1.16.0
tzdata==2024.1
urllib3==2.2.2
